Implement rudimentary interface
Clients can send pixels to port 1234 via TCP.
The server interprets every seven bytes as one pixel, this means that
every binary file is a valid stream of pixels.
All integers are read in network byteorder and coordinates start in the
top left corner.
The first two bytes represent the x coordinate, the next two bytes the y
coordinate. The next three bytes are interpret as the red, green and
blue value of the pixel.